KeySelection splitted into Parsing and Validation model
previously the pydantic KeySelection model was created each time the select_relevant_keys function was called. To optimize this a the model was seperated into an own file and split into a pydantic model for llm response parsing and a model for the post-response validation. Neccessary due to the impossibility to provide dynamically assigned context to the parse function of the OpenAI Client lib